This script sends a roster reminder
This script takes variables from roster_reminder_sample.ini including a roster view number. 
An email is generated for each participant in the roster view - which includes either a table or a list of tasks for this week.
A second email is sent to the nominated roster coordinator - informing them of anyone who does not have and email address in Jethro and so will not have received the roster reminder.